requirements.txt
atlassian-python-api==3.41.11
email_validator==2.1.1
flask==3.0.3
flask_bcrypt==1.0.1
flask_login==0.6.3
flask_mail==0.9.1
flask_migrate==4.0.7
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
google-generativeai==0.5.4
gunicorn==22.0.0
influxdb-client==1.43.0
jinja2==3.1.4
openai==1.30.1
Pillow==10.3.0
pydantic==2.7.1
python-decouple==3.8
pyyaml==6.0.1
reportlab==4.2.0
requests==2.31.0
sqlalchemy==2.0.30
Werkzeug==3.0.3
portalocker==2.8.2